app/requirements.txt
flask
flask-sqlalchemy
flask_marshmallow
marshmallow-sqlalchemy
mysql-python
python-dotenv
